version 2

+++ src/user/services/user.service.ts
@@ -1,8 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { User } from 'src/user/user.entity';
+import { User } from '../user.entity';
 import { Repository } from 'typeorm';
-import { UserSignupDto } from './dto/user-signup.dto';
-import { stringToHash } from 'src/helpers/string.helpers';
 import { InjectRepository } from '@nestjs/typeorm';
 
 @Injectable()
@@ -11,41 +9,47 @@ export class UserService {
     @InjectRepository(User) private readonly repository: Repository<User>,
   ) {}
 
-  create(dto: UserSignupDto) {
+  create(email: string, password: string): Promise<User> {
     const user = this.repository.create({
-      email: dto.email,
-      password: stringToHash(dto.password),
+      email,
+      password,
     });
 
     return this.repository.save(user);
   }
 
-  find(email?: string) {
+  find(email?: string): Promise<User[]> {
     return this.repository.find({
       where: { email },
     });
   }
 
-  async findOne(id: number) {
-    const user = await this.repository.findOne({ where: { id } });
+  findOne(id: number): Promise<User> {
+    if (!id) {
+      return null;
+    }
+
+    return this.repository.findOne({ where: { id } });
+  }
+
+  async update(id: number, attributes: Partial<User>): Promise<User> {
+    const user = await this.findOne(id);
 
     if (!user) {
       throw new NotFoundException('User not found!');
     }
-
-    return user;
-  }
-
-  async update(id: number, attributes: Partial<User>) {
-    const user = await this.findOne(id);
 
     Object.assign(user, attributes);
 
     return this.repository.save(user);
   }
 
-  async remove(id: number) {
+  async remove(id: number): Promise<User> {
     const user = await this.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found!');
+    }
 
     return this.repository.remove(user);
   }
